Bug fixes found during 5.0.3backporting of features
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@6121 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/extensions/i2/org.wcs.smart.i2/src/org/wcs/smart/i2/xml/XmlToIntelData.java
@@ -218,14 +218,16 @@ public class XmlToIntelData {
 			if (e.getBirtTemplate() != null) {
 				Path src = rootPath.resolve(e.getBirtTemplate());
 				Path trg = IntelReportManager.INSTANCE.getEntityTemplate(e);
-				try {
-					if (!Files.exists(trg.getParent())) {
-						Files.createDirectories(trg.getParent());
+				if(Files.exists(src)) {
+					try {
+						if (!Files.exists(trg.getParent())) {
+							Files.createDirectories(trg.getParent());
+						}
+						Files.copy(src, trg, StandardCopyOption.REPLACE_EXISTING);
+					} catch (IOException e1) {
+						copyErrors.add(MessageFormat.format(Messages.XmlToIntelData_EntityCopyError, e.getName()));
+						Intelligence2PlugIn.log(e1.getMessage(), e1);
 					}
-					Files.copy(src, trg, StandardCopyOption.REPLACE_EXISTING);
-				} catch (IOException e1) {
-					copyErrors.add(MessageFormat.format(Messages.XmlToIntelData_EntityCopyError, e.getName()));
-					Intelligence2PlugIn.log(e1.getMessage(), e1);
 				}
 			}
 		});
